tetridiscobot/pyproject.toml
[tool.poetry]
name = "tetridiscobot"
version = "0.1.0"
description = ""
authors = ["Dylann Batisse <dylannbatisse@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.0.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
